fix code convert column
